+++ index.html
@@ -228,6 +228,6 @@
     </div>
   </template>
   
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
